fix(tooltip): bind correct 'hide' event handler
If a user defines a tooltip-trigger attribute, this ensures the correct event handlers is used to hide the tooltip.

Fixes a bug where if a user sets both a default trigger using the tooltip provider, and then tries to override with an attribute, the wrong 'hide' event was being used.